Fixed Message Hider
