+++ SkyblockUtilsV2/data/Settings.js
@@ -45,7 +45,7 @@ class Settings {
   @SwitchProperty({
     name: "Vanquisher Spawn Message",
     description: "Send a Message in Party Chat containing your location if a Vanquisher Spawns",
-    category: "Nether",
+    category: "Nether"
   })
   vanquisherMessage = false;
 
@@ -53,22 +53,24 @@ class Settings {
   @SwitchProperty({
     name: "Phrase Filter",
     description: "Enable/Disable a filter for certain Phrases or Words from a list you make using /su word-list add myWord",
-    category: "Filter",
+    category: "Filter"
   })
-  phrasesFilter = false;
+  phraseFilter = false;
 
   //Waypoints
   @SwitchProperty({
     name: "Waypoint shown",
     description: "If Waypoints you add should be shown in-game",
-    category: "Waypoint",
+    category: "Waypoint"
   })
-  waypointShown = false
+  waypointShown = false;
 
   @SwitchProperty({
     name: "Auto Waypoint Pickup",
-    description: "If Waypoints detected in chat should be shown automatically (Only works if Waypoint shown is enabled)
+    description: "If Waypoints detected in chat should be shown automatically (Only works if Waypoint shown is enabled)",
+    category: "Waypoint"
   })
+  autoWaypoint = false;
 
   constructor() {
     this.initialize(this);
