+++ SkyblockUtilsV2/utils/fileUtils.js
@@ -1,4 +1,3 @@
-import * as constants from "./constants";
 import * as Const from "./constants";
 
 export function list(file) {
